Making each field in the target and inner structure be an array of 1, 2, 4, or 8 byte entries. This change will allow a mapping between a target field and inner field only if the arrays corresponding to these fields have the same number of entries. New symbolic variables field<N>_n indicate the number of entries in the array corresponding to adapted inner field <N>.
Another change is to make the memory layout adaptor search by
parameterized by 2 parameters: (1) the number of fields to be mapped (2)
the size of the structure in bytes of the target or inner structure,
whichever is larger.
Adding FuzzBALL implementation to provide these parameters via FuzzBALL
options and changing the top-level Perl script similarly.