Structure formula gen. stats, and switch to hash-based unique
Print information about the time spent creating formulas for
structure adaptors, currently for the AS-mode code only because of
duplicaion.
Switch from an O(n^2) unique function to a O(n) one based on a hash
table: this is a major speedup but to what is not the most expensive
part of the process.