Adding table treatment for symbolic regions.
SRFM#region_expr returns 5 value now in a tuple, the 3rd value being a
bool which indicates if the region expression treatment can be handled
using a table.
If the table treatment is being applied, then the 2nd value (addr) will
be set to 0 and if it is not being applied, then the 5th value (cloc)
will be set to 0.