Side-conditions + End-of-f1 SE eq. checking Adding side conditions so that for every field (end_byte - start_byte + 1)/f_n should be either 1, 2, 4, or 8. Also, f_type should equal only valid values.
Once f1 finishes, side-effects of f1 should be adapted to f2 only once
for each entry in opt_synth_struct_adaptor list.
During the adaptor search step, this assumes that
addresses in opt_synth_struct_adaptor will be in the same order as that
in which they appear in the list of test cases.

This change also enforces f1's initial memory to be adapted to f2 once
only per run of f2.

This End-of-f1 SE eq checking change gets rid of a Double unsat bug that
shows up when doing adaptor search on more than one test case,
the side-effects of the first run of f1 are not equivalent to the second
run of f1.