Check path condition before fm#start_symbolic
Lack of the stat system call causes a binary dynamically loading C
libraries to encounter symbolic branches during the loading on X86-64.
The fix for this issue is two-pronged:
(1) Add support for the stat system call for X86-64
(2) stop fuzzball if the path condition is non-empty before we get a
chance to run fm#start_symbolic, which takes snapshots of memory and
processes extra conditions (esp. important for adaptor synthesis)