Fixes to run ARM executables with FuzzBALL
Used arm-linux-gnueabi-gcc version 4.7.3 to cross compile executables to
ARM.
This includes the following fixes:
(1) (fix in fragment_machine.ml#eval_int_exp_ty) If-Then-Else loads will not evaluate both, the Then and the Else,
sides. We will first check if the ITE condition is concrete, and then
only evaluate the side the ITE condition evaluates to. This is to
prevent noop instructions which are loads from causing null
dereferences.
(2) (fix in fragment_machine.ml#run_sl) utores with If-Then-Else rhs expressions which are noops should be treated as noops by
checking if the lhs expression is the same as the rhs expression that
the ITE condition on the rhs evaluates to.
(3) (fix in libasmir/src/vex/irtoir-arm.cpp) Registers such as CCDEP1 were showing up in the execution of ARM
instructions such as tstne. This register (along with CCDEP2, CCNDEP, CCOP) are Valgrind's scratch
registers which Valgrind uses to lazily evaluate flags. FuzzBALL wants
eager evaluation of flags, and so, tries to not have these show up in
its Vine IR translation. To fix this, we changed the default direction
we choose when we search the VEX IR statement list for the latest update
to the flags. We also changed the default direction where the updated
value appears when constructing the ITE expression into ASM IR.
(4) (fix in fragment_machine.ml#make_arm_regs_zero) R_D0 - R_D31 were not part of make_arm_regs_zero causing them also
to show up in ARM instruction execution with uninitialized values, even though they are not
supposed to. We just set their initial values to 0.