Instead of using collapse_temps in formula_manager#simplify_exp, use the simplified expanded version of an expression in formula_manager#simplify_exp only if the simplified expanded version of the expression is smaller than the expression.
collapse_temps uses subexpr_to_temp_var_info which is maintained
across FuzzBALL iterations. This means that a subexpression that
is made equal to a temp in one FuzzBALL iteration will be
collapsed into that temp in later iterations. This can affect caches in
FuzzBALL that perform syntactic equivalence checking of
expressions (like SRFM.concrete_cache) and cause a decision tree inconsistency
because the cache isn't populated with the same entries across FuzzBALL iterations