dont allow pure symbolic regions during adapter search
The actual changes in sym_region_frag_machine.ml were made in the
previous commit. But I decided to add a comment and create a trail of
this change in git commits.

We dont want to allow symbolic regions to be used during adapter search
because even though an adapter may be found under some assumption of
symbolic region usage by the target and reference functions, these
symbolic regions aren't going to be part of any adapter family.