Use a more appropriate bound for concrete memory address chosen for equality with an expression for which a region is being created in SRFM#region_for.
When FuzzBALL chooses to make a symbolic region for an expression, it
first tries to use concrete memory addresses existing within the
expression. If it can find such concrete memory addresses within the
expression, it adds a constraint in the path condition that the concrete
address is equal to the region expression and uses concrete memory at
that address instead of using a symbolic region.

But, for figuring out which concrete addresses are appropriate to be
mined from within the region expression, FuzzBALL will now check if the
mined concrete address lies within the bound (starting_sane_addr,
    starting_sane_addr + 10000 * !opt_region_limit) where
starting_sane_addr is set to the constant value in SRFM#region_for (and
    also synth-argsub.pl, synth-typeconv.pl, synth-arithmetic.ml) and
10000 is the maximum number of steps expected for CE search (since
    symbolic regions are only used within CE search steps).