Turn on retsub repair adapter + add verify adapter options
1. This commit turns on the repair return value substitution adaptor but
disables substitution of the return value by one of the callee arguments
2. This commit adds a few options to make FuzzBALL run in a
"verify-adapter" mode. In this mode, FuzzBALL will run the target
fragment only once with concrete adapter given as input. It will report
a "Mismatch" if a violation of the check-condition-at is found. This
should make the wrapper script (currently only synth-argsub-repair.pl)
  move to running try_synth once again.